USB Camera Stream
This version of this project implements a simple USB Webcam stream. At a rate of twenty times per second, a frame is retrieved, and processed using OpenCV filters. This program will be expended in the future to calculate x, y, and z positions and velocities, as well as grab frames from an Axis IP Camera.