Real-time Computer Vision
This version of this project opens a camera stream from an Axis m1013 IP Camera with a static IP Address connected to a raspberry pi over wi-fi. A frame is pulled from the camera 15 times per second, and the frame is processed using OpenCV filters. A VisionState is communicated between the raspberry pi and roborio that are both also connected via wi-fi. This Vision State is used by the pi for distance measurement purposes.