Switcheable Object Tracking
This version of this project employs a UDP server to communicate between roborio and a raspberry pi. The client UDP Socket receives a vision state, which directs the raspberry pi which vision target to open and process. This application dynamically loads vision tracking preferences and thresholds from a text file of the raspberry pi.