Incidence Angle
This version of this project fixes a number of bugs with the UDP Client, and calculates a horizontal incidence angle with the vision target rather than x and y positions. This calculation is used to incrementally aim towards the vision target. This application publishes an angle to the UDP server to be read by a NI RoboRIO device.